requirements.txt
torch
torchvision
onnx
onnxruntime
numpy
pillow
matplotlib 
ultralytics